refactorizacion completa

+++ src/modules/users/users.controller.ts
@@ -1,15 +1,17 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, ForbiddenException, Req } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { AuthGuard } from '../auth/auth.guard';
+import { Public } from '../auth/decorators/public.decorators';
 
 @Controller('users')
 @UseGuards(AuthGuard)
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(private readonly usersService: UsersService) { }
 
-  @Post()
+  @Post('register')
+  @Public()
   create(@Body() createUserDto: CreateUserDto) {
     return this.usersService.register(createUserDto);
   }
@@ -20,12 +22,21 @@ export class UsersController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: number, @Body() updateUserDto: UpdateUserDto) {
+  update(@Param('id') id: number, @Body() updateUserDto: UpdateUserDto, @Req() req: any) {
+
+    const userRole = req.user?.role;
+    if (userRole !== 'ADMIN') {
+      throw new ForbiddenException('Only ADMIN can update users');
+    }
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: number) {
+  remove(@Param('id') id: number, @Req() req: any) {
+    const userRole = req.user?.role;
+    if (userRole !== 'ADMIN') {
+      throw new ForbiddenException('Only ADMIN can delete users');
+    }
     return this.usersService.remove(id);
   }
 }
