+++ src/modules/users/users.service.ts
@@ -11,7 +11,7 @@ export class UsersService {
 
   async register(createUserDto: CreateUserDto) {
     try {
-      const { password, username, role, plantaId } = createUserDto;
+      const { password, username, role, plantId } = createUserDto;
 
       const existingUser = await this.prisma.user.findUnique({
         where: { username },
@@ -23,17 +23,24 @@ export class UsersService {
 
       const hash = await bcrypt.hash(password, 10);
 
-      return this.prisma.user.create({
+      const newUser = await this.prisma.user.create({
         data: {
           username,
           password: hash,
           role,
-          plantaId, 
+          plantId,
         },
       });
 
+      return {
+        id: newUser.id,
+        username: newUser.username,
+        role: newUser.role,
+        plantId: newUser.plantId ? newUser.plantId : null,
+      };
+
     } catch (error) {
-      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
+      throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
@@ -52,24 +59,45 @@ export class UsersService {
 
     } catch (error) {
 
-      return new HttpException(error.message, HttpStatus.BAD_REQUEST)
+      return new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR)
     }
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
 
     try {
+      const { password, username, role, plantId } = updateUserDto;
+      const existingUser = await this.prisma.user.findUnique({
+        where: { id },
+      });
+      const existingPlant = plantId ? await this.prisma.plant.findUnique({ where: { id: plantId } }) : null;
+      if (plantId && !existingPlant) {
+        throw new HttpException('Plant not found', HttpStatus.NOT_FOUND);
+      }
+      if (!existingUser) {
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+      }
 
       const updatedUser = await this.prisma.user.update({
         where: { id },
-        data: updateUserDto,
-      });
+        data: {
+          username: username ? username : existingUser.username,
+          role: role ? role : existingUser.role,
+          plantId: plantId ? plantId : existingUser.plantId,
+          password: password ? await bcrypt.hash(password, 10) : existingUser.password,
+        }
+      })
 
-      return updatedUser
+      return {
+        id: updatedUser.id,
+        username: updatedUser.username,
+        role: updatedUser.role,
+        plantId: updatedUser.plantId ? updatedUser.plantId : null,
+      }
 
     } catch (error) {
 
-      return new HttpException(error.message, HttpStatus.BAD_REQUEST)
+      return new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR)
     }
 
   }
@@ -77,6 +105,12 @@ export class UsersService {
   async remove(id: number) {
 
     try {
+      const existingUser = await this.prisma.user.findUnique({
+        where: { id },
+      });
+      if (!existingUser) {
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+      }
 
       const removedUser = await this.prisma.user.delete({
         where: { id },
